requirements.txt
python-telegram-bot==20.7
paramiko==3.4.0
requests==2.31.0
psutil==5.9.8
pyyaml==6.0.1
